+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2023/06/25
+
+### Added
+
+- New classes for custom event types
+  - `CollisionEvent`
+  - `CollisionEvent2D`
+  - `ColliderEvent`
+  - `ColliderEvent2D`
+- New `layerMask` property added to all event behaviors
+- Help URLs added to behaviors
+
+### Changed
+
+- Set namespace of all events to `Zigurous.Physics.Events`
+
 ## [0.3.1] - 2021/06/28
 
 ### Changed